Extracted CoronaTestQRCode handling
